feat(protocol-designer): batch edit mix form
