+++ protocol-designer/src/components/BatchEditForm/BatchEditMix.js
@@ -1,0 +1,208 @@
+// @flow
+import * as React from 'react'
+import {
+  Box,
+  PrimaryButton,
+  OutlineButton,
+  Tooltip,
+  useHoverTooltip,
+  TOOLTIP_TOP,
+  TOOLTIP_FIXED,
+} from '@opentrons/components'
+import { i18n } from '../../localization'
+import {
+  BlowoutLocationField,
+  CheckboxRowField,
+  DelayFields,
+  FlowRateField,
+  TipPositionField,
+  WellOrderField,
+} from '../StepEditForm/fields'
+import {
+  getBlowoutLocationOptionsForForm,
+  getLabwareFieldForPositioningField,
+} from '../StepEditForm/utils'
+import { FormColumn } from './FormColumn'
+import type { FieldPropsByName } from '../StepEditForm/types'
+import type { WellOrderOption } from '../../form-types'
+// TODO(IL, 2021-03-01): refactor these fragmented style rules (see #7402)
+import formStyles from '../forms/forms.css'
+import styles from '../StepEditForm/StepEditForm.css'
+import buttonStyles from '../StepEditForm/ButtonRow/styles.css'
+
+type BatchEditMixProps = {|
+  batchEditFormHasChanges: boolean,
+  propsForFields: FieldPropsByName,
+  handleCancel: () => mixed,
+  handleSave: () => mixed,
+|}
+export const BatchEditMix = (props: BatchEditMixProps): React.Node => {
+  const { propsForFields, handleCancel, handleSave } = props
+  const [cancelButtonTargetProps, cancelButtonTooltipProps] = useHoverTooltip({
+    placement: TOOLTIP_TOP,
+    strategy: TOOLTIP_FIXED,
+  })
+  const [saveButtonTargetProps, saveButtonTooltipProps] = useHoverTooltip({
+    placement: TOOLTIP_TOP,
+    strategy: TOOLTIP_FIXED,
+  })
+  const disableSave = !props.batchEditFormHasChanges
+
+  const getLabwareIdForPositioningField = (name: string): string | null => {
+    const labwareField = getLabwareFieldForPositioningField(name)
+    const labwareId = propsForFields[labwareField]?.value
+    return labwareId ? String(labwareId) : null
+  }
+
+  const getPipetteIdForForm = (): string | null => {
+    const pipetteId = propsForFields.pipette?.value
+    return pipetteId ? String(pipetteId) : null
+  }
+
+  const getWellOrderFieldValue = (name: string): ?WellOrderOption => {
+    const val = propsForFields[name]?.value
+    if (val === 'l2r' || val === 'r2l' || val === 't2b' || val === 'b2t') {
+      return val
+    } else {
+      return null
+    }
+  }
+
+  return (
+    <div className={formStyles.form}>
+      <Box className={styles.form_wrapper}>
+        <Box className={styles.section_wrapper}>
+          <FormColumn
+            sectionHeader={i18n.t('form.batch_edit_form.settings_for', {
+              prefix: 'aspirate',
+            })}
+          >
+            <Box className={styles.form_row}>
+              <FlowRateField
+                {...propsForFields['aspirate_flowRate']}
+                pipetteId={getPipetteIdForForm()}
+                flowRateType="aspirate"
+              />
+              <TipPositionField
+                {...propsForFields['mix_mmFromBottom']}
+                labwareId={getLabwareIdForPositioningField('mix_mmFromBottom')}
+              />
+              <WellOrderField
+                prefix="mix"
+                label={i18n.t('form.step_edit_form.field.well_order.label')}
+                firstValue={getWellOrderFieldValue('mix_wellOrder_first')}
+                secondValue={getWellOrderFieldValue('mix_wellOrder_second')}
+                firstName="mix_wellOrder_first"
+                secondName="mix_wellOrder_second"
+                updateFirstWellOrder={
+                  propsForFields['mix_wellOrder_first'].updateValue
+                }
+                updateSecondWellOrder={
+                  propsForFields['mix_wellOrder_second'].updateValue
+                }
+              />
+            </Box>
+
+            <DelayFields
+              checkboxFieldName={'aspirate_delay_checkbox'}
+              secondsFieldName={'aspirate_delay_seconds'}
+              labwareId={getLabwareIdForPositioningField(
+                'aspirate_delay_mmFromBottom'
+              )}
+              propsForFields={propsForFields}
+            />
+          </FormColumn>
+
+          <FormColumn
+            sectionHeader={i18n.t('form.batch_edit_form.settings_for', {
+              prefix: 'dispense',
+            })}
+          >
+            <FlowRateField
+              {...propsForFields['dispense_flowRate']}
+              pipetteId={getPipetteIdForForm()}
+              flowRateType="dispense"
+            />
+            <DelayFields
+              checkboxFieldName={'dispense_delay_checkbox'}
+              secondsFieldName={'dispense_delay_seconds'}
+              labwareId={getLabwareIdForPositioningField(
+                'dispense_delay_mmFromBottom'
+              )}
+              propsForFields={propsForFields}
+            />
+            <CheckboxRowField
+              {...propsForFields['mix_touchTip_checkbox']}
+              label={i18n.t('form.step_edit_form.field.touchTip.label')}
+              className={styles.small_field}
+            >
+              <TipPositionField
+                {...propsForFields['mix_touchTip_mmFromBottom']}
+                labwareId={getLabwareIdForPositioningField(
+                  'mix_touchTip_mmFromBottom'
+                )}
+              />
+            </CheckboxRowField>
+            <CheckboxRowField
+              {...propsForFields['blowout_checkbox']}
+              label={i18n.t('form.step_edit_form.field.blowout.label')}
+              className={styles.small_field}
+            >
+              {/* TODO(IL, 2021-03-03): Location is supported only with same-pipette, right? Which options? */}
+              <BlowoutLocationField
+                {...propsForFields['blowout_location']}
+                className={styles.full_width}
+                options={getBlowoutLocationOptionsForForm({
+                  stepType: 'mix',
+                })}
+              />
+            </CheckboxRowField>
+          </FormColumn>
+          {/* <SourceDestBatchEditMoveLiquidFields
+            prefix="aspirate"
+            propsForFields={propsForFields}
+          />
+          <SourceDestBatchEditMoveLiquidFields
+            prefix="dispense"
+            propsForFields={propsForFields}
+          /> */}
+        </Box>
+
+        <Box textAlign="right" maxWidth="55rem" marginTop="3rem">
+          <Box
+            {...cancelButtonTargetProps}
+            marginRight="0.625rem"
+            display="inline-block"
+          >
+            <OutlineButton
+              onClick={handleCancel}
+              className={buttonStyles.button_auto}
+            >
+              {i18n.t('button.discard_changes')}
+            </OutlineButton>
+            <Tooltip {...cancelButtonTooltipProps}>
+              {i18n.t('tooltip.cancel_batch_edit')}
+            </Tooltip>
+          </Box>
+
+          <Box
+            {...saveButtonTargetProps}
+            className={buttonStyles.form_button}
+            display="inline-block"
+          >
+            <PrimaryButton disabled={disableSave} onClick={handleSave}>
+              {i18n.t('button.save')}
+            </PrimaryButton>
+            <Tooltip {...saveButtonTooltipProps}>
+              {i18n.t(
+                `tooltip.save_batch_edit.${
+                  disableSave ? 'disabled' : 'enabled'
+                }`
+              )}
+            </Tooltip>
+          </Box>
+        </Box>
+      </Box>
+    </div>
+  )
+}
